\name{yacca-package}
\alias{yacca-package}
\alias{yacca}
\docType{package}
\title{
Yet Another Canonical Correlation Analysis Package
}
\description{
This package provides an alternative canonical correlation/redundancy analysis function, with associated print, plot, and summary methods.  A method for generating helio plots is also included.
}
\details{

For details on using the package, see \code{\link{cca}} and \code{\link{helio.plot}}.
}
\author{
Carter T. Butts <buttsc@uci.edu>

Maintainer: Carter T. Butts <buttsc@uci.edu>
}
\references{
Mardia, K. V.; Kent, J. T.; and Bibby, J. M.  1979.  \emph{Multivariate Analysis}.  London: Academic Press.
}

\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
